fix: error display
